mvp_cold_chain/requirements.txt
# Core Data Science
pandas==2.1.0
numpy==1.25.0
scipy==1.11.0

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.16.0

# Interactive Maps
folium==0.14.0
branca==0.6.0

# Geospatial Analysis
geopandas==0.13.2
shapely==2.0.1
geopy==2.3.0

# API & Web
requests==2.31.0
urllib3==2.0.4
python-dotenv==1.0.0

# Data Processing
openpyxl==3.1.2  # Excel files
xlrd==2.0.1      # Old Excel format

# Model Interpretation
shap==0.42.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Progress Bars
tqdm==4.66.1

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Web App
streamlit==1.28.0
